feat : login and signup functionality

+++ src/pages/NotePage.jsx
@@ -2,62 +2,85 @@ import React , {useState,useEffect} from 'react'
 import { useParams , useNavigate} from 'react-router-dom'
 import {ReactComponent as ArrowLeft} from '../assests/arrow-left.svg'
 
-const NotePage = () => {
+const NotePage = (props) => {
     let {id} = useParams();
+    let isLoggedIn = props.isLoggedIn;
     let [note,setNote] = useState(null);
+  
+    
     const navigate = useNavigate();
+
+
+    let apiHost =  'http://127.0.0.1:8000' 
+
+    const headers = {
+        'Content-Type': 'application/json',
+        Authorization: `Bearer ${localStorage.getItem('access_token')}`,
+      };
     
     useEffect( () => {
-        getNote()
+        if(isLoggedIn){
+            getNote()    
+        }
     },[id])
 
-    let apiHost =  'https://jam-notes-be.vercel.app' 
 
     let getNote = async () => {
         
-        let response = await fetch(apiHost + `/api/note/${id}`)
+        const options = {
+            method: 'GET',
+            headers: headers,
+          };
+
+        let response = await fetch(apiHost + `/api/note/${id}`,options)
         let data = await response.json()
     
         setNote(data)
     }
 
     let updateNote = async () =>{
-        await fetch(apiHost + `/api/note/${id}/`,{
-            method:'PATCH',
-            headers:{
-                "Content-type":'application/json'
-            },
+        const options = {
+            method: 'PATCH',
+            headers: headers,
             body:JSON.stringify({
                 "title":note.title,
                 "text":note.text
             })
-        })
+          };
+
+        await fetch(apiHost + `/api/note/${id}/`,options)
     }
     
     let deleteNote = async () => {
-        await fetch(apiHost + `/api/note/${id}/`,{
-            method:'DELETE',
-          
-        })
+        if (isLoggedIn){
+            const options = {
+                method: 'DELETE',
+                headers: headers,
+            };
+            await fetch(apiHost + `/api/note/${id}/`,options)
+        }
         navigate(-1);
     }
+
     let handleSubmit = () =>{
-   
-        if (note.title === '' && note.text === '' ){
-            deleteNote()
+        if (isLoggedIn){
+            
+            if (note.title === '' && note.text === '' ){
+                deleteNote()
+            }else {     
+                updateNote()         
+            }               
         }
-        else {
-            updateNote()
-            navigate(-1);
-        }
+        navigate(-1);      
     }
     
     return <>
-            <div className='note'>
+
+            <div className='note container my-3'>
                 <div className='d-flex'>
                     <input type="text" defaultValue={note?.title} onInput={(e) =>{setNote({...note,'title':e.target.value})}} placeholder='title'/>
                     <h1>
-                        <ArrowLeft className='arrow-left' onClick={handleSubmit} />
+                        <ArrowLeft className='arrow-left' onClick={handleSubmit}  />
                     </h1>
                 </div>
                 <textarea onInput={(e) =>{setNote({...note,'text':e.target.value})}} defaultValue={note?.text} placeholder='your note'></textarea>
@@ -65,7 +88,7 @@ const NotePage = () => {
                 <button type="button" className="btn btn-outline-danger" onClick={deleteNote} >Delete</button>
 
             </div>
-                
+     
         </>
 }
